\name{data_nosnp}
\alias{data_nosnp}
\docType{data}
\title{
Simulated covariate data
}
\description{
Covariate data including outcome (binary), sex, age and 18 PCs for 20,000 individuals.
}
\usage{data("data_nosnp")}
\format{
  A data frame with 20,000 observations on a binary outcome named \code{outcome}
  and 20 numeric covariates named \code{sex}, \code{age} and \code{PC1}-\code{PC18}.
}
\examples{
data(data_nosnp)
}
\keyword{datasets}
