\name{example_data_nosnp}
\alias{example_data_nosnp}
\docType{data}
\title{
Example covariate data
}
\description{
Example covariate data including outcome (binary), sex, age and 18 PCs for 5,000 individuals.
}
\usage{data("example_data_nosnp")}
\format{
  A data frame with 5,000 observations on a binary outcome named \code{outcome}
  and 20 numeric covariates named \code{sex}, \code{age} and \code{PC1}-\code{PC18}.
}
\examples{
data(example_data_nosnp)
}
\keyword{datasets}
