\name{example_hap2}
\alias{example_hap2}
\docType{data}
\title{Example genotype data for the second genome}
\description{Example genotype data for the second genome of 8 SNPs for 5,000 individuals.}
\usage{data("example_hap2")}
\format{A data frame with 5,000 observations on 8 binary variables named \code{SNP1}-\code{SNP8}
  ("0" denotes the reference allele while "1" denotes the alternative allele).}
\examples{
data(example_hap2)
}
\keyword{datasets}
