% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_htrx.R
\name{make_htrx}
\alias{make_htrx}
\alias{make_htr}
\alias{make_snp}
\title{Generate haplotype data}
\usage{
make_htrx(
  SNP1,
  SNP2,
  rareremove = FALSE,
  rare_threshold = 0.001,
  fixedfeature = NULL
)

make_htr(SNP1, SNP2, rareremove = FALSE, rare_threshold = 0.001)

make_snp(SNP1, SNP2, rareremove = FALSE, rare_threshold = 0.001)
}
\arguments{
\item{SNP1}{a data frame of the SNPs' genotype of the first genome. The genotype of a SNP for each individual is either 0 (reference allele) or 1 (alternative allele).}

\item{SNP2}{a data frame of the SNPs' genotype of the second genome. The genotype of a SNP for each individual is either 0 (reference allele) or 1 (alternative allele).}

\item{rareremove}{logical. Remove rare SNPs and haplotypes or not. By default, rareremove=FALSE.}

\item{rare_threshold}{a numeric number below which the haplotype or SNP is removed. This only works when rareremove=TRUE.}

\item{fixedfeature}{a character consisted of the names of haplotypes.}
}
\value{
a data frame of the feature data (either for SNPs, HTR or HTRX).
}
\description{
Generate the feature data, either the genotype data for SNPs,
the feature data for Haplotype Trend Regression (HTR), or
the feature data for Haplotype Trend Regression with eXtra flexibility (HTRX).
}
\details{
If there are n SNPs, there are \ifelse{html}{\out{2<sup>n</sup>}}{\eqn{2^n}} different haplotypes created by HTR,
and \ifelse{html}{\out{3<sup>n</sup>}}{\eqn{3^n}}-1 different haplotypes created by HTRX.
}
\examples{
## create SNP data for both genomes
SNP1=as.data.frame(matrix(0,nrow=100,ncol=4))
SNP2=as.data.frame(matrix(0,nrow=100,ncol=4))
colnames(SNP1)=colnames(SNP2)=c('a','b','c','d')
p=runif(4,0.01,0.99)
for(j in 1:4){
  SNP1[,j]=rbinom(n,1,p[j])
  SNP2[,j]=rbinom(n,1,p[j])
}

## create the SNP data without removing rare SNPs
make_snp(SNP1,SNP2)

## create feature data for "HTR" removing haplotypes rarer than 0.5\%
make_htr(SNP1,SNP2,TRUE,0.005)

## create feature data for "HTRX" without removing haplotypes
make_htrx(SNP1,SNP2)

## create feature data for feature"01XX" and "X101" without removing haplotypes
make_htrx(SNP1,SNP2,fixedfeature=c("01XX","X101"))
}
