\name{SNP1}
\alias{SNP1}
\docType{data}
\title{Simulated genotype data for genome 1}
\description{Simulated genotype data for the first genome of 8 SNPs for 20,000 individuals.}
\usage{data("SNP1")}
\format{A data frame with 20,000 observations on 8 binary variables named \code{SNP1}-\code{SNP8}
  ("0" denotes the reference allele while "1" denotes the alternative allele).}
\examples{
data(SNP1)
}
\keyword{datasets}


