% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themodel.R
\name{themodel}
\alias{themodel}
\title{Model fitting}
\usage{
themodel(formula, data = NULL, usebinary = 1, clean = TRUE, ...)
}
\arguments{
\item{formula}{a formula for model-fitting.}

\item{data}{a data frame contains all the variables included in the formula.
The phenotype variable must be named "outcome".}

\item{usebinary}{a non-negative number representing different models.
Use linear model if usebinary=0, use logistic regression model via fastglm if usebinary=1,
and use logistic regression model via glm if usebinary>1.}

\item{clean}{logical. If clean=TRUE (by default), remove additional storages that
the "predict", "AIC", "BIC" methods do not need.}
}
\value{
a fitted model.
}
\description{
Model-agnostic functions for model fitting (both linear and generalized linear models).
}
\details{
"fasglm" is from "fastglm" package, which is much faster than "glm".
}
\examples{
## create the dataset for variables and outcome
x=matrix(runif(100,-2,2),ncol=5)
outcome=0.5*x[,2] - 0.8*x[,4] + 0.3*x[,5]
data1=data.frame(outcome,x)

## fit a linear model
themodel(outcome~.,data1,usebinary=0)

## create binary outcome
outcome=outcome>runif(100,-2,2)
outcome[outcome]=1
data2=data.frame(outcome,x)

## fit a logistic regression model
themodel(outcome~.,data2,usebinary=1)
}
